+++ src/BulletCollision/BroadphaseCollision/btDbvt.h
@@ -1249,7 +1249,7 @@ inline void btDbvt::rayTestInternal(const btDbvtNode* root,
 			bounds[1] = node->volume.Maxs() - aabbMin;
 			btScalar tmin = 1.f, lambda_min = 0.f;
 			unsigned int result1 = false;
-			result1 = btRayAabb2(rayFrom, rayDirectionInverse, signs, bounds, tmin, lambda_min, lambda_max);
+			result1 = (unsigned int)btRayAabb2(rayFrom, rayDirectionInverse, signs, bounds, tmin, lambda_min, lambda_max);
 			if (result1)
 			{
 				if (node->isinternal())
@@ -1289,7 +1289,7 @@ inline void btDbvt::rayTest(const btDbvtNode* root,
 		rayDirectionInverse[0] = rayDir[0] == btScalar(0.0) ? btScalar(BT_LARGE_FLOAT) : btScalar(1.0) / rayDir[0];
 		rayDirectionInverse[1] = rayDir[1] == btScalar(0.0) ? btScalar(BT_LARGE_FLOAT) : btScalar(1.0) / rayDir[1];
 		rayDirectionInverse[2] = rayDir[2] == btScalar(0.0) ? btScalar(BT_LARGE_FLOAT) : btScalar(1.0) / rayDir[2];
-		unsigned int signs[3] = {rayDirectionInverse[0] < btScalar(0.0), rayDirectionInverse[1] < btScalar(0.0), rayDirectionInverse[2] < btScalar(0.0)};
+		unsigned int signs[3] = {(unsigned int)(rayDirectionInverse[0] < btScalar(0.0)), (unsigned int)(rayDirectionInverse[1] < btScalar(0.0)), (unsigned int)(rayDirectionInverse[2] < btScalar(0.0))};
 
 		btScalar lambda_max = rayDir.dot(rayTo - rayFrom);
 
@@ -1316,7 +1316,7 @@ inline void btDbvt::rayTest(const btDbvtNode* root,
 			bounds[1] = node->volume.Maxs();
 
 			btScalar tmin = 1.f, lambda_min = 0.f;
-			unsigned int result1 = btRayAabb2(rayFrom, rayDirectionInverse, signs, bounds, tmin, lambda_min, lambda_max);
+			unsigned int result1 = (unsigned int)btRayAabb2(rayFrom, rayDirectionInverse, signs, bounds, tmin, lambda_min, lambda_max);
 
 #ifdef COMPARE_BTRAY_AABB2
 			btScalar param = 1.f;
